Mais ajustes

+++ paisagem.cpp
@@ -3,8 +3,8 @@
 #include <R.h>
 #include <Rmath.h>
 
-paisagem::paisagem(double raio, int N, double angulo_visada, double passo, double move, double taxa_basal, 
-				   double taxa_morte, double incl_b, double incl_d, int numb_cells, double cell_size, int land_shape, 
+paisagem::paisagem(double raio, int N, double angulo_visada, double passo, double move, double taxa_basal,
+				   double taxa_morte, double incl_b, double incl_d, int numb_cells, double cell_size, int land_shape,
 				   int density_type, double death_mat, int inipos, int bound_condition, int scape[]):
 	tamanho(numb_cells*cell_size),
 	N(N),
@@ -14,7 +14,7 @@ paisagem::paisagem(double raio, int N, double angulo_visada, double passo, doubl
 	landscape_shape(land_shape),
 	boundary_condition(bound_condition),
 	initialPos(inipos)
-{	
+{
 		for(int i=0;i<this->numb_cells;i++)
 		{
 			for(int j=0;j<this->numb_cells;j++)
@@ -23,33 +23,33 @@ paisagem::paisagem(double raio, int N, double angulo_visada, double passo, doubl
 				this->landscape[i][j]=scape[j*numb_cells+i];
 			}
 		}
-		
+
 	//Atribui valores à matriz patches, que determina o fragmento a que cada pixel pertence
 	int component = 0;
-	for (int k = 0; k < this->numb_cells; ++k) 
-		for (int l = 0; l < this->numb_cells; ++l) 
+	for (int k = 0; k < this->numb_cells; ++k)
+		for (int l = 0; l < this->numb_cells; ++l)
 			if (!this->patches[k][l] && this->landscape[k][l]) find_patches(k, l, ++component);
 	this->numb_patches = component;
-		
-	this->patch_area = new int[this->numb_patches+1];
-		
+
+	this->patch_area = new double[this->numb_patches+1];
+
 	for (unsigned int j = 0; j<numb_patches+1; j++)
 		this->patch_area[j] = 0;
-	
+
 	for(int i = 0; i < this->numb_cells; i++)
 		for(int j = 0; j < this->numb_cells; j++)
 			this->patch_area[patches[i][j]] += 1;
 	for (unsigned int j = 0; j<numb_patches+1; j++)
 		this->patch_area[j] = this->patch_area[j]*this->cell_size*this->cell_size;
-		
+
 	// Calculo do raio dependendo do tipo de densidade. 0 = global, 1 = local (raio), 2 = kernel.
 	if(density_type==0)
 	{
 		raio = this->tamanho/sqrt(M_PI);
 	}
-	/* Coloca os indivíduos na paisagem por meio da função populating() */	
+	/* Coloca os indivíduos na paisagem por meio da função populating() */
 	this->populating(raio,N,angulo_visada,passo,move,taxa_basal,taxa_morte,incl_b,incl_d,death_mat,density_type);
-	
+
 	for(unsigned int i=0; i<this->popIndividuos.size(); i++)
 	{
 		this->atualiza_vizinhos(this->popIndividuos[i]);//atualiza os vizinhos
@@ -63,9 +63,9 @@ paisagem::paisagem(double raio, int N, double angulo_visada, double passo, doubl
 		double dsty=this->calcDensity(popIndividuos[i]);
 		this->popIndividuos[i]->update(dsty);   //e atualiza o individuo i da populacao
 	}
-	
+
 }
-		
+
 void paisagem::populating(double raio, int N, double angulo_visada, double passo, double move, double taxa_basal,
 						  double taxa_morte, double incl_b, double incl_d, double death_m,
 						  int dens_type)
@@ -116,7 +116,7 @@ void paisagem::populating(double raio, int N, double angulo_visada, double passo
 														incl_d,
 														death_m,
 														dens_type));
-		}	
+		}
     }
 	if(this->initialPos==2) // Random initial positions with normal distribution. TBI: tornar os parametros da rnorm livres
 	{
@@ -139,13 +139,13 @@ void paisagem::populating(double raio, int N, double angulo_visada, double passo
 														death_m,
 														dens_type));
 		}
-	}	
+	}
 }
 
 int paisagem::update()
 {
     if(this->popIndividuos.size()>0)
-    {    
+    {
 	// Este for loop pode ser paralelizado, pois o que acontece com cada individuo eh independente
 	#ifdef PARALLEL
 	#pragma omp parallel for
@@ -163,7 +163,7 @@ int paisagem::update()
 			double dsty=this->calcDensity(popIndividuos[i]);
             this->popIndividuos[i]->update(dsty);   //e atualiza o individuo i da populacao
         }
-		
+
 	}
 }
 
@@ -172,7 +172,7 @@ int paisagem::sorteia_individuo()
 	// time for next event and simulation time update
 	int menor=0;
 	double menor_tempo = this->popIndividuos[0]->get_tempo();
-	
+
 	for(unsigned int i=1; i<this->popIndividuos.size(); i++)
 	{
 		if(this->popIndividuos[i]->get_tempo()<menor_tempo)
@@ -201,7 +201,7 @@ bool paisagem::realiza_acao(int acao, int lower) //TODO : criar matriz de distan
         this->popIndividuos.push_back(chosen);
         break;
 
-    case 2: 
+    case 2:
         this->popIndividuos[lower]->anda();
 	emigrou = this->apply_boundary(popIndividuos[lower]);
 		break;
@@ -242,7 +242,7 @@ bool paisagem::apply_boundary(individuo * const ind) //const
 	double rad = (double)ind->get_raio();
 	switch(this->boundary_condition)
 	{
-			
+
 		case 0:
 		if(this->landscape_shape==0)
 		{
@@ -259,7 +259,7 @@ bool paisagem::apply_boundary(individuo * const ind) //const
 				emigrou = true;
 			}
 		}
-		
+
 		if(this->landscape_shape==1)
 		{
 			if((double)ind->get_x()>=this->numb_cells*this->cell_size/2 || //>= porque na paisagem quadrado as bordas mais distantes de 0 iniciariam um proximo pixel que estaria fora da paisagem. Ou teriamos que assumir que esses pixels mais extremos tenha uma área maior, o que daria um trabalho adicional para implementar uma situação irreal.
@@ -276,10 +276,10 @@ bool paisagem::apply_boundary(individuo * const ind) //const
 					}
 				}
 				emigrou = true;
-			}			   
-		}		
+			}
+		}
 		break;
-	
+
 		case 1:
 		if(ind->get_x()<(this->numb_cells*this->cell_size/2)*(-1))
 			ind->set_x(this->tamanho+ind->get_x());
@@ -305,7 +305,7 @@ double paisagem::calcDist(const individuo* a1, const individuo* a2) const //Viro
 		case 0:
 			return sqrt(((double)a1->get_x()-(double)a2->get_x())*((double)a1->get_x()-(double)a2->get_x())+((double)a1->get_y()-(double)a2->get_y())*((double)a1->get_y()-(double)a2->get_y()));
 			break;
-	
+
 		case 1:
 			double x1=a1->get_x();
 			double x2=a2->get_x();
@@ -321,19 +321,19 @@ double paisagem::calcDist(const individuo* a1, const individuo* a2) const //Viro
 }
 
 // A function to calculate de density of individuals according to density type (global or local) and considering landscape boundary effects in the calculation.
-double paisagem::calcDensity(const individuo* ind1) const 
+double paisagem::calcDensity(const individuo* ind1) const
 {
 	double density;
 	density=ind1->NBHood_size()/(M_PI*(ind1->get_raio()*ind1->get_raio()));
-	
-	// Functions for local density calculation 
-	
+
+	// Functions for local density calculation
+
 	/* 1. Circular area defining a region in which denso-dependence occurs: landscape boundary effects.
-	 In this case, density is the number of individuals inside the circle divided by circle area.  
+	 In this case, density is the number of individuals inside the circle divided by circle area.
 	 This is the same calculation as for global density, except by the cases in which landscape boundary affects
-	 the area of the circle.			
+	 the area of the circle.
 	 */
-	
+
 	// Condition giving the boundary effects cases
 	if(ind1->get_densType()==1)
 	{
@@ -346,7 +346,7 @@ double paisagem::calcDensity(const individuo* ind1) const
 			double XYmax = this->tamanho/2;
 			vector<double>secX;
 			vector<double>secY;
-			
+
 			// Functions for adjusted local density calculation, according to the specific case
 			// 1)
 			if(modIx>XYmax-ind1->get_raio() && modIy<=XYmax-ind1->get_raio())
@@ -355,16 +355,16 @@ double paisagem::calcDensity(const individuo* ind1) const
 				secX.push_back(XYmax);
 				secY.push_back(modIy+sqrt(ind1->get_raio()*ind1->get_raio()-((XYmax-modIx)*(XYmax-modIx))));
 				secY.push_back(modIy-sqrt(ind1->get_raio()*ind1->get_raio()-((XYmax-modIx)*(XYmax-modIx))));
-				
+
 				double distSecs = secY[0]-secY[1];
 				double height = XYmax - modIx;
 				double theta = acos(1-(distSecs*distSecs/(2*ind1->get_raio()*ind1->get_raio()))); // angle in radians
 				double adjArea = M_PI*ind1->get_raio()*ind1->get_raio() - (theta*ind1->get_raio()*ind1->get_raio()/2 - (distSecs*height/2));
-				
+
 				density = ind1->NBHood_size()/adjArea;
-				
+
 			}
-			
+
 			// 2)
 			if(modIx<=XYmax-ind1->get_raio() && modIy>XYmax-ind1->get_raio())
 			{
@@ -372,19 +372,19 @@ double paisagem::calcDensity(const individuo* ind1) const
 				secY.push_back(XYmax);
 				secX.push_back(modIx+sqrt(ind1->get_raio()*ind1->get_raio()-((XYmax-modIy)*(XYmax-modIy))));
 				secX.push_back(modIx-sqrt(ind1->get_raio()*ind1->get_raio()-((XYmax-modIy)*(XYmax-modIy))));
-				
+
 				double distSecs = secX[0]-secX[1];
 				double height = XYmax - modIy;
 				double theta = acos(1-(distSecs*distSecs/(2*ind1->get_raio()*ind1->get_raio()))); // angle in radians
 				double adjArea = M_PI*ind1->get_raio()*ind1->get_raio() - (theta*ind1->get_raio()*ind1->get_raio()/2 - (distSecs*height/2));
-				
+
 				density = ind1->NBHood_size()/adjArea;
 			}
-			
-			
+
+
 			if(modIx>XYmax-ind1->get_raio() && modIy>XYmax-ind1->get_raio())
 			{
-				
+
 				// 3)
 				if((modIx-XYmax)*(modIx-XYmax)+(modIy-XYmax)*(modIy-XYmax)>ind1->get_raio()*ind1->get_raio())
 				{
@@ -396,41 +396,41 @@ double paisagem::calcDensity(const individuo* ind1) const
 					secY.push_back(modIy+sqrt(ind1->get_raio()*ind1->get_raio()-((XYmax-modIx)*(XYmax-modIx))));
 					secX.push_back(XYmax);
 					secY.push_back(modIy-sqrt(ind1->get_raio()*ind1->get_raio()-((XYmax-modIx)*(XYmax-modIx))));
-					
+
 					double distSecs = sqrt((secX[3]-secX[1])*(secX[3]-secX[1])+(secY[1]-secY[3])*(secY[1]-secY[3]));
 					double distSecs2 = sqrt((secX[2]-secX[0])*(secX[2]-secX[0])+(secY[0]-secY[2])*(secY[0]-secY[2]));
 					double theta = acos(1-(distSecs*distSecs/(2*ind1->get_raio()*ind1->get_raio()))); // angle in radians
 					double phi = acos(1-(distSecs2*distSecs2/(2*ind1->get_raio()*ind1->get_raio()))); // angle in radians
 					double adjArea = (2*M_PI-theta)*ind1->get_raio()*ind1->get_raio()/2 + phi*ind1->get_raio()*ind1->get_raio()/2 + (secX[0]-secX[1])*(XYmax-modIy)/2 + (secY[2]-secY[3])*(XYmax-modIx)/2;
-					
+
 					density = ind1->NBHood_size()/adjArea;
-					
+
 				}
 				// 4)
-				else 
+				else
 				{
 					secX.push_back(modIx-sqrt(ind1->get_raio()*ind1->get_raio()-((XYmax-modIy)*(XYmax-modIy))));
 					secY.push_back(XYmax);
 					secX.push_back(XYmax);
 					secY.push_back(modIy-sqrt(ind1->get_raio()*ind1->get_raio()-((XYmax-modIx)*(XYmax-modIx))));
-					
+
 					double distSecs = sqrt((secX[1]-secX[0])*(secX[1]-secX[0])+(secY[0]-secY[1])*(secY[0]-secY[1]));
 					double theta = acos(1-(distSecs*distSecs/(2*ind1->get_raio()*ind1->get_raio()))); // angle in radians
 					double adjArea = theta*ind1->get_raio()*ind1->get_raio()/2 + (XYmax-secX[0])*(XYmax-modIy) + (XYmax-secY[1])*(XYmax-modIx);
-					
+
 					density = ind1->NBHood_size()/adjArea;
 				}
-				
+
 			}
 		}
 	}
-	
+
 	/* 2. Density kernel (TBI).
-	 
+
 	 if(ind1->get_densType()==2) {}
-	 
+
 	 */
-	return density;	
+	return density;
 }
 
 
@@ -440,14 +440,14 @@ double paisagem::calcDensity(const individuo* ind1) const
 */
 
 void paisagem::atualiza_vizinhos(individuo * const ag1) const //acessando os vizinhos dos agentes
-{  
+{
 	vector <individuo*> listViz;
 	if(ag1->get_densType()==0) //dens_type poderia voltar como propriedade da paisagem. Facilitariam as coisas. Como muitas propriedades e métodos deste código, elas podem ser interpretadas das duas formas (como do individuo ou como da paisagem). O que está dando confusão é que estamos fazendo um IBM, mas para algumas situações estamos querendo simular dinâmicas cujas variáveis de interesse são propriedades populacionais e não do indivíduo. Se aceito, limar o método get_densType() do individuo.h.
-	{		
+	{
 		for(unsigned int j=0; j<popIndividuos.size(); j++)
 		{
 			individuo* ag2=this->popIndividuos[j];
-			if(ag1==ag2) continue;			   
+			if(ag1==ag2) continue;
 			listViz.push_back(ag2);
 		}
 	}
@@ -457,23 +457,23 @@ void paisagem::atualiza_vizinhos(individuo * const ag1) const //acessando os viz
 		for(unsigned int j=0; j<popIndividuos.size(); j++)
 		{
 			individuo* ag2=this->popIndividuos[j];
-			if(ag1==ag2) continue;   
+			if(ag1==ag2) continue;
 			double d=this->calcDist(ag1,ag2);
-			if(d<=rad) {listViz.push_back(ag2);}		
+			if(d<=rad) {listViz.push_back(ag2);}
 		}
 	}
 	ag1->set_vizinhos(listViz);
-		
+
 }
 
 void paisagem::atualiza_habitat(individuo * const ind) const
 {
-	// Tinha um IF com landscape_shape que eu removi. Não entendi como a paisagem ser circular 
+	// Tinha um IF com landscape_shape que eu removi. Não entendi como a paisagem ser circular
 	// interfere em ser habitat ou não: isso deve interferir na apply_boundary apenas, certo?
 	// Also: Tinha uma inversão do y que eu também não entendi e removi
 	// A.C. 10.07.13
-	
-	// Um termo (-1) foi removido erroneamente por A.C.. Para o hy, o sentido em que o número de células aumenta é o 
+
+	// Um termo (-1) foi removido erroneamente por A.C.. Para o hy, o sentido em que o número de células aumenta é o
 	//contrário do sentido em que as coordenadas aumentam. Portanto a multiplicação por - 1 é necessária.
 	// M.A. 12.09.14
 	int hx,hy;
